Acceso_a_Datos/PR9 - Acceso por Voz/Asistente_Miriam - Kasa/requirements.txt
psycopg2-binary
SpeechRecognition
pyaudio
python-dotenv